size_t in for loops

+++ src/KDTree.cpp
@@ -21,7 +21,7 @@ KDTree::~KDTree() {
 KDTree* KDTree::createTree(const std::vector<Point3d>& pointsOrig) {
     std::vector<Point3d>::size_type const n = pointsOrig.size();
     std::vector<const Point3d*> points(n);
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         points[i] = &pointsOrig[i];
     }
     return createTree(points, points.begin(), points.end(), 0);
@@ -30,7 +30,7 @@ KDTree* KDTree::createTree(const std::vector<Point3d>& pointsOrig) {
 KDTree* KDTree::createTree(const std::vector<Point3d*>& pointsOrig) {
     std::vector<Point3d*>::size_type const n = pointsOrig.size();
     std::vector<const Point3d*> points(n);
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         points[i] = pointsOrig[i];
     }
     return createTree(points, points.begin(), points.end(), 0);
